Move variants-to-JSON to utils
